+++ tests/test_service_details.py
@@ -3,22 +3,66 @@ import os
 
 from pytest import raises
 
-from idf_component_manager.service_details import NoSuchProfile, service_details
+from idf_component_manager.service_details import NoSuchProfile, get_namespace, get_profile, get_token, service_details
+from idf_component_tools.config import ConfigManager
 from idf_component_tools.errors import FatalError
 
 
 class TestServiceDetails(object):
-    def test_service_details_without_namespace(self):
-        with raises(FatalError, match='Namespace*'):
-            service_details()
+    def test_get_namespace_without_namespace(self):
+        config = ConfigManager().load().profiles.get('default', {})
+        with raises(FatalError, match='Failed to get namespace*'):
+            get_namespace(config)
 
-    def test_service_details_without_token(self):
-        with raises(FatalError, match='API token*'):
-            service_details(namespace='test')
+    def test_get_namespace_with_namespace(self):
+        config = ConfigManager().load().profiles.get('default', {})
+        namespace = get_namespace(config, 'example')
+        assert namespace == 'example'
 
-    def test_service_details_without_profile(self):
-        with raises(NoSuchProfile, match='"test" didn\'t find*'):
-            service_details(service_profile='test', namespace='test')
+    def test_get_token(self):
+        config = ConfigManager().load().profiles.get('default', {})
+        with raises(FatalError, match='Failed to get API Token*'):
+            get_token(config)
+
+    def test_get_profile_success(self, tmp_path):
+        config_path = os.path.join(str(tmp_path), 'idf_component_manager.yml')
+        with open(config_path, 'w+') as f:
+            f.write(
+                json.dumps(
+                    {
+                        'profiles': {
+                            'test': {
+                                'url': 'https://example.com/',
+                                'default_namespace': 'espressif',
+                                'api_token': 'token'
+                            }
+                        }
+                    }))
+        profile = get_profile(config_path, 'test')
+        assert profile['url'] == 'https://example.com/'
+        assert profile['default_namespace'] == 'espressif'
+        assert profile['api_token'] == 'token'
+
+    def test_get_profile_not_exist(self, tmp_path):
+        config_path = os.path.join(str(tmp_path), 'idf_component_manager.yml')
+        with open(config_path, 'w+') as f:
+            f.write(
+                json.dumps(
+                    {
+                        'profiles': {
+                            'test': {
+                                'url': 'https://example.com/',
+                                'default_namespace': 'espressif',
+                                'api_token': 'token'
+                            }
+                        }
+                    }))
+
+        assert get_profile(config_path, 'not_test') == {}
+
+    def test_service_details_namespace_not_exist(self):
+        with raises(NoSuchProfile):
+            service_details(service_profile='not_exists')
 
     def test_service_details_success(self, tmp_path):
         with open(os.path.join(str(tmp_path), 'idf_component_manager.yml'), 'w+') as f: